# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Character API
    description: HTTP service for managing characters.
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /items:
        get:
            tags:
                - item
            summary: getItems item
            operationId: item#getItems
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ItemResponse'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ItemGetItemsBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ItemGetItemsIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/ItemGetItemsIDsExhaustedResponseBody'
            schemes:
                - http
        post:
            tags:
                - item
            summary: createItem item
            operationId: item#createItem
            parameters:
                - name: CreateItemRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ItemCreateItemRequestBody'
                    required:
                        - name
                        - description
                        - damage
                        - healing
                        - protection
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ItemCreateItemResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - damage
                            - healing
                            - protection
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ItemCreateItemBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ItemCreateItemIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/ItemCreateItemIDsExhaustedResponseBody'
            schemes:
                - http
    /items/{id}:
        get:
            tags:
                - item
            summary: getItem item
            operationId: item#getItem
            parameters:
                - name: id
                  in: path
                  description: Item ID
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ItemGetItemResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - damage
                            - healing
                            - protection
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ItemGetItemBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ItemGetItemIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/ItemGetItemIDsExhaustedResponseBody'
            schemes:
                - http
        put:
            tags:
                - item
            summary: updateItem item
            operationId: item#updateItem
            parameters:
                - name: id
                  in: path
                  description: Item ID
                  required: true
                  type: integer
                - name: UpdateItemRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ItemUpdateItemRequestBody'
                    required:
                        - name
                        - description
                        - damage
                        - healing
                        - protection
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ItemUpdateItemResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - damage
                            - healing
                            - protection
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ItemUpdateItemBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ItemUpdateItemIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/ItemUpdateItemIDsExhaustedResponseBody'
            schemes:
                - http
        delete:
            tags:
                - item
            summary: deleteItem item
            operationId: item#deleteItem
            parameters:
                - name: id
                  in: path
                  description: Item ID
                  required: true
                  type: integer
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ItemDeleteItemBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/ItemDeleteItemIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/ItemDeleteItemIDsExhaustedResponseBody'
            schemes:
                - http
definitions:
...
    ItemCreateItemBadDataResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createItem_BadData_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemCreateItemIDNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: createItem_IdNotFound_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemCreateItemIDsExhaustedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createItem_IDsExhausted_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemCreateItemRequestBody:
        title: ItemCreateItemRequestBody
        type: object
        properties:
            damage:
                type: integer
                description: Item damage
                example: 3866523682817493181
                format: int64
            description:
                type: string
                description: Item description
                example: Illum possimus est maxime suscipit.
            healing:
                type: integer
                description: Item healing
                example: 15809512099163135118
                format: int64
            name:
                type: string
                description: Item name
                example: Deserunt in eos hic assumenda.
            protection:
                type: integer
                description: Item protection
                example: 3286063385544471345
                format: int64
        example:
            damage: 15206467097202700645
            description: Amet accusamus voluptas quod.
            healing: 16773944419522702605
            name: Sit beatae velit dicta tempora.
            protection: 8802686892147945185
        required:
            - name
            - description
            - damage
            - healing
            - protection
    ItemCreateItemResponseBody:
        title: ItemCreateItemResponseBody
        type: object
        properties:
            damage:
                type: integer
                description: Item damage
                example: 11428886412422101053
                format: int64
            description:
                type: string
                description: Item description
                example: Sit impedit aperiam dolorem.
            healing:
                type: integer
                description: Item healing
                example: 17629313834758374553
                format: int64
            id:
                type: integer
                description: Item ID
                example: 5941048070598964254
                format: int64
            name:
                type: string
                description: Item name
                example: Impedit consequatur perspiciatis.
            protection:
                type: integer
                description: Item protection
                example: 8706003594395890363
                format: int64
        example:
            damage: 17175827231517026164
            description: Saepe voluptatem facilis.
            healing: 3932466711290375790
            id: 1034414631917739976
            name: Unde accusantium.
            protection: 961445790002134063
        required:
            - id
            - name
            - description
            - damage
            - healing
            - protection
...
    ItemGetItemsBadDataResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: getItems_BadData_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemGetItemsIDNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: getItems_IdNotFound_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemGetItemsIDsExhaustedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: getItems_IDsExhausted_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    ItemResponse:
        title: ItemResponse
        type: object
        properties:
            damage:
                type: integer
                description: Item damage
                example: 453487864614646226
                format: int64
            description:
                type: string
                description: Item description
                example: Officiis placeat dolorem deserunt.
            healing:
                type: integer
                description: Item healing
                example: 7335432543203285654
                format: int64
            id:
                type: integer
                description: Item ID
                example: 132768975321920730
                format: int64
            name:
                type: string
                description: Item name
                example: Temporibus illum delectus debitis.
            protection:
                type: integer
                description: Item protection
                example: 5125630491184283553
                format: int64
        description: Item and attributes
        example:
            damage: 3409983216479890797
            description: Ut sed sed temporibus.
            healing: 2711073850677007024
            id: 3489247875915087112
            name: Tenetur est dolorem cumque dolorem velit voluptatem.
            protection: 7300025597621537829
        required:
            - id
            - name
            - description
            - damage
            - healing
            - protection
...